refactor(vm): 优化虚拟机寄存器和调用帧的数据类型
- 将 CallFrame 中的返回值类型从 Opt<u64> 改为 Opt<u16>
- 将 CallFrame 中的 _sp 类型从 u64 改为 u32
- 将 Chunk 中异常值寄存器类型从 u64 改为 u16
- 将 FastRegisterPool 中指针访问参数类型从 size_t 改为 u64
- 将 VMState 中寄存器访问方法参数类型从 u64 改为 u16
- 将 IRGenerator 中寄存器索引类型从 u64 改为 u16
- 添加寄存器数量溢出检查防止超出 u16 最大值
- 修改 VMState 中 reg 方法实现方式，区分值获取和引用获取
- 将局部变量添加方法参数从右值引用改为值传递
- 修正闭包创建逻辑中的栈深度判断条件

+++ src/vm/Chunk.hpp
@@ -26,7 +26,7 @@ namespace cial::Bytecode {
     struct ThrowHandler {
         u64 tryStart;
         u64 tryEnd;
-        u64 exValueReg;
+        u16 exValueReg;
     };
 
     class Chunk : public MarkSweepHeader {
